Small fix for type of variable

+++ lib/screens/location_screen.dart
@@ -34,7 +34,7 @@ class _LocationScreenState extends State<LocationScreen> {
         cityName = '';
         return;
       }
-      double temp = weatherData['main']['temp'];
+      double temp = weatherData['main']['temp'].toDouble();
       temperature = temp.toInt();
       var condition = weatherData['weather'][0]['id'];
       cityName = weatherData['name'];
